requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
grad-cam>=1.4.0
matplotlib>=3.5.0
opencv-python>=4.5.0
